add better label handling
